상품등록/수정 완료(파일처리까지.......다함..........)

+++ src/main/java/com/example/monthly/service/board/ProductService.java
@@ -56,14 +56,7 @@ public class ProductService {
         }
 
     }
-//    상품삭제
-    public void removeProduct(Long productNumber){
-        if(productNumber== null){
-            throw new IllegalArgumentException("상품번호 누락");
-        }
-        productFileService.remove(productNumber);
-        productMapper.deleteProduct(productNumber);
-    }
+
 //    상품수정
     public void modifyProduct(ProductDto productDto){
         if(productDto == null){
@@ -72,14 +65,29 @@ public class ProductService {
         productMapper.updateProduct(productDto);
     }
 
-    public void modifyProduct(ProductDto productDto, List<MultipartFile>files, MultipartFile file) throws IOException {
+    public void modifyProduct(ProductDto productDto, List<MultipartFile>files, MultipartFile file){
         if(productDto==null || files==null || file==null){
             throw new IllegalArgumentException("제품 수정 매개변수 null체크");
         }
         Long productNumber = productDto.getProductNumber();
+        System.out.println(productNumber);
+        if(!files.isEmpty()){
         productFileService.remove(productNumber); //원래있던 파일은 지우고
-        productFileService.registerAndSaveMainFile(file, "m",productNumber); //대표사진다시저장
-        productFileService.registerAndSaveFiles(files, "d",productNumber); //리스트다시저장
+            try {
+                productFileService.registerAndSaveFiles(files, "d",productNumber); //리스트다시저장
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        if(!file.isEmpty()){
+            productFileService.removeMainFile(productNumber);
+            try {
+                productFileService.registerAndSaveMainFile(file, "m",productNumber); //대표사진다시저장
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
+        }
         productMapper.updateProduct(productDto);
     }
